+++ template/js/admin.js
@@ -4,4 +4,4 @@ import './netlify-cms/pages-preview'
 
 document.title = `Admin ~ ${document.title}`
 
-initNetlifyCms()
+initNetlifyCms(window.CMS_CUSTOM_CONFIG)