+++ template/js/checkout.js
@@ -1,9 +1,7 @@
 import './lib/config'
 import './lib/fetch-info'
+import './lib/utils'
 
 import '@ecomplus/storefront-app'
-
-import './lib/icons'
-import './lib/utils'
 
 import(/* webpackPrefetch: true */ './lib/load-widgets')